account for distance, & slitely different fitting
I added an argument to specify the standoff distance, since that may occasionally matter (the previus calculation implicitly and incorrectly assumed a distance of 1μm).

perhaps more importantly, I changed the definition of the error bars for the least squares to go like sqrt(n) instead of (incorrectly) sqrt(1/n).  this makes fits that look much better and give what I think are more reasonable temperatures (tho the fact that it makes a significant difference is a little concerning).

I also modified the residuals plot to be relative residuals, so the y axis is more meaningful.